Restrict actions to admin

+++ app/controllers/chefs_controller.rb
@@ -4,6 +4,8 @@ class ChefsController < ApplicationController
 
   before_action :require_same_user, only: [:edit, :update, :destroy]
   
+  before_action :require_admin, only: [:destroy]
+  
   def index
     @chefs = Chef.paginate(page: params[:page], per_page: 2)
   end
@@ -40,12 +42,14 @@ class ChefsController < ApplicationController
   end
   
   def destroy
-    chefname = @chef.chefname
-    if @chef.destroy
-      flash[:danger] = "Chef '#{chefname}' was deleted, along with all associated recipes!"
-      redirect_to chefs_path
-    else
-      render "show"
+    if !@chef.admin?
+      chefname = @chef.chefname
+      if @chef.destroy
+        flash[:danger] = "Chef '#{chefname}' was deleted, along with all associated recipes!"
+        redirect_to chefs_path
+      else
+        render "show"
+      end
     end
   end
   
@@ -61,9 +65,16 @@ class ChefsController < ApplicationController
     end
 
     def require_same_user
-      if current_chef != @chef
+      if current_chef != @chef && !current_chef.admin
         flash[:danger] = "You can only edit or delete your own account!"
         redirect_to chefs_path
+      end
+    end
+    
+    def require_admin
+      if logged_in? && !current_chef.admin?
+        flash[:danger] = "Only administrators can perform that action!"
+        redirect_to root_path
       end
     end
 end
